refactor(website): split graph benchmark components
