+++ website/src/content/docs/benchmark/index.mdx
@@ -1,5 +1,4 @@
-import GraphBenchmark from "../../../components/benchmark/GraphBenchmark";
-import PerformanceSummary from "../../../components/benchmark/PerformanceSummary";
+import TtscWebsiteBenchmarkGraphCommon from "../../../components/benchmark/graph/TtscWebsiteBenchmarkGraphCommon";
 
 # Benchmark
 
@@ -11,7 +10,7 @@ import PerformanceSummary from "../../../components/benchmark/PerformanceSummary
 
 The overview below uses the same shared onboarding prompt across every project with Codex GPT mini. The detailed page follows with model tabs, project-specific prompts, and structural coverage.
 
-<GraphBenchmark variant="summary" />
+<TtscWebsiteBenchmarkGraphCommon summary />
 
 ## Compiler Performance
