+++ website/src/content/docs/graph/design.mdx
@@ -1,6 +1,6 @@
 import { Tabs } from "nextra/components";
 
-import LocalSource from "../../../components/LocalSource";
+import TtscWebsiteLocalSource from "../../../components/TtscWebsiteLocalSource";
 
 # Design
 
@@ -194,67 +194,67 @@ These tabs read from `packages/graph/src/structures`, so the reference stays tie
   ]}
 >
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphNext.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphEntrypoints.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphLookup.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphTrace.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphTour.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphDetails.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphOverview.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphEscape.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphNode.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphEdge.ts"
       showLineNumbers
     />
   </Tabs.Tab>
   <Tabs.Tab>
-    <LocalSource
+    <TtscWebsiteLocalSource
       path="packages/graph/src/structures/ITtscGraphDiagnostic.ts"
       showLineNumbers
     />
