+++ website/src/content/docs/graph/index.mdx
@@ -1,4 +1,4 @@
-import GraphBenchmark from "../../../components/benchmark/GraphBenchmark";
+import TtscWebsiteBenchmarkGraphCommon from "../../../components/benchmark/graph/TtscWebsiteBenchmarkGraphCommon";
 
 # Code Graph (MCP)
 
@@ -22,7 +22,7 @@ The graph hands it the answer up front through one MCP tool, `inspect_typescript
 
 Here is the median token cost on the shared onboarding question, `@ttsc/graph` against no MCP and the three comparators. Lower is better, and it stays flat as the repository grows.
 
-<GraphBenchmark variant="summary" />
+<TtscWebsiteBenchmarkGraphCommon summary />
 
 The [full benchmark](/docs/benchmark/graph) has every model, the per-repository prompts, the structural coverage, and the method.
