+++ website/src/content/docs/graph/viewer.mdx
@@ -1,4 +1,4 @@
-import GraphViewer3D from "../../../components/graph/GraphViewer3D";
+import TtscWebsiteGraphViewer3D from "../../../components/graph/TtscWebsiteGraphViewer3D";
 
 # 3D Graph Viewer
 
@@ -6,7 +6,7 @@ The same checker-resolved graph the MCP tools answer from, rendered as a navigab
 
 Drag to orbit, scroll to zoom, hover a node for its name, kind, and file.
 
-<GraphViewer3D />
+<TtscWebsiteGraphViewer3D />
 
 ## 1. Use it on your own project
