+++ website/src/content/index.mdx
@@ -2,6 +2,6 @@
 title: ttsc, a TypeScript-Go toolchain
 ---
 
-import TtscLandingMovie from "../movies/landing/TtscLandingMovie";
+import TtscWebsiteLandingMovie from "../movies/landing/TtscWebsiteLandingMovie";
 
-<TtscLandingMovie />
+<TtscWebsiteLandingMovie />
