+++ website/src/movies/landing/TtscWebsiteLandingFadeIn.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+interface TtscWebsiteLandingFadeInProps {
+  children: React.ReactNode;
+  delay?: number;
+  className?: string;
+}
+
+export default function TtscWebsiteLandingFadeIn({
+  children,
+  className = "",
+}: TtscWebsiteLandingFadeInProps) {
+  return <div className={className}>{children}</div>;
+}
