+++ website/src/movies/landing/TtscWebsiteLandingInTheBrowser.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import FadeIn from "./FadeIn";
-import SectionEyebrow from "./SectionEyebrow";
+import TtscWebsiteLandingFadeIn from "./TtscWebsiteLandingFadeIn";
+import TtscWebsiteLandingSectionEyebrow from "./TtscWebsiteLandingSectionEyebrow";
 
 const STEPS = [
   "edit TypeScript",
@@ -10,12 +10,12 @@ const STEPS = [
   "compare output",
 ] as const;
 
-export default function InTheBrowser() {
+export default function TtscWebsiteLandingInTheBrowser() {
   return (
     <section className="relative overflow-hidden bg-neutral-950 px-6 py-24 md:py-32">
       <div className="relative mx-auto max-w-6xl">
-        <FadeIn>
-          <SectionEyebrow label="Playground" />
+        <TtscWebsiteLandingFadeIn>
+          <TtscWebsiteLandingSectionEyebrow label="Playground" />
           <div className="grid gap-10 lg:grid-cols-[1fr_0.9fr] lg:items-center">
             <div>
               <h2 className="text-3xl font-bold leading-[1.08] tracking-tight text-white md:text-5xl">
@@ -72,7 +72,7 @@ export default function InTheBrowser() {
               </div>
             </div>
           </div>
-        </FadeIn>
+        </TtscWebsiteLandingFadeIn>
       </div>
     </section>
   );
