+++ website/src/movies/landing/TtscWebsiteLandingSponsors.tsx
@@ -1,20 +1,20 @@
 "use client";
 
-import FadeIn from "./FadeIn";
-import SectionEyebrow from "./SectionEyebrow";
+import TtscWebsiteLandingFadeIn from "./TtscWebsiteLandingFadeIn";
+import TtscWebsiteLandingSectionEyebrow from "./TtscWebsiteLandingSectionEyebrow";
 
 const SPONSORS_IMAGE =
   "https://raw.githubusercontent.com/samchon/sponsor-images/refs/heads/master/public/circle.svg";
 const SPONSORS_URL = "https://github.com/sponsors/samchon";
 
-export default function Sponsors() {
+export default function TtscWebsiteLandingSponsors() {
   return (
     <section className="relative overflow-hidden border-t border-neutral-900 bg-neutral-950 px-6 py-24 md:py-32">
       <div className="absolute inset-x-0 top-0 h-px bg-gradient-to-r from-transparent via-cyan-300/40 to-transparent" />
 
       <div className="relative mx-auto grid max-w-6xl gap-10 lg:grid-cols-[0.9fr_1.1fr] lg:items-center">
-        <FadeIn>
-          <SectionEyebrow label="Sponsors" />
+        <TtscWebsiteLandingFadeIn>
+          <TtscWebsiteLandingSectionEyebrow label="Sponsors" />
           <h2 className="max-w-xl text-3xl font-bold leading-[1.08] tracking-tight text-white md:text-5xl">
             Built with help from people who keep the work moving.
           </h2>
@@ -36,9 +36,9 @@ export default function Sponsors() {
               Thank you for your support
             </p>
           </div>
-        </FadeIn>
+        </TtscWebsiteLandingFadeIn>
 
-        <FadeIn delay={120}>
+        <TtscWebsiteLandingFadeIn delay={120}>
           <a
             href={SPONSORS_URL}
             target="_blank"
@@ -53,7 +53,7 @@ export default function Sponsors() {
               loading="lazy"
             />
           </a>
-        </FadeIn>
+        </TtscWebsiteLandingFadeIn>
       </div>
     </section>
   );
